requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
anthropic==0.21.0
httpx==0.25.2
pytest==8.3.0
pytest-asyncio==0.24.0
python-dotenv==1.0.0
bcrypt==4.2.0
PyJWT==2.10.0
asyncio-mqtt==0.16.2
fastapi-users==12.1.3
starlette==0.27.0
sqlalchemy==2.0.23
stripe==8.7.0
